% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cal_coverage}
\alias{cal_coverage}
\title{calculate genome coverage}
\usage{
cal_coverage(loc_df)
}
\arguments{
\item{loc_df}{data.frame. must contain 4 columns: \code{chrom}, \code{strand}, \code{start}, \code{end}}
}
\value{
tibble. 4 elements, \code{chrom}, \code{strand}, \code{pos}, \code{n_reads}. \code{n_reads} means number of reads covering given position.
}
\description{
calculate genome coverage
}
\details{
we use 1-based coordinate, \link{start, end}
}
\examples{
sam_file <- system.file('extdata', 'Neat1_1.Aligend_trunc.sam', package = 'paristools')
sam_file \%>\% read_sam() \%>\% sam_to_loc_df() \%>\% cal_coverage()

}
