% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_coverage_impl}
\alias{cal_coverage_impl}
\title{calculate genome coverage for one chromosome}
\arguments{
\item{chrom_loc_df}{data.frame. Locations on one chromosome, columns are \code{strand}, \code{start}, \code{end}. There must be at least one region of minimum length 1.}
}
\value{
tibble. Columns are \code{pos}, \code{+}, \code{-}, the latter two are integers representing the genome, index is 1-based genome position, value is number of reads covering that position.
}
\description{
calculate genome coverage for one chromosome
}
\details{
we use 1-based coordinate, \link{start, end}
}
\keyword{internal}
