% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_dg_coverage}
\alias{get_dg_coverage}
\title{calculate average coverage of duplex group segment}
\usage{
get_dg_coverage(duplex_group, genome_coverage)
}
\arguments{
\item{duplex_group}{tibble. see \code{\link[=read_duplexgroup]{read_duplexgroup()}}}

\item{genome_coverage}{tibble. see \code{\link[=cal_coverage]{cal_coverage()}}}
}
\value{
tibble.
}
\description{
calculate average coverage of duplex group segment
}
\examples{
sam_file <- system.file('extdata', 'Neat1_1.Aligend_trunc.sam', package = 'paristools')
duplexgroup_file <- system.file('extdata', 'Neat1_1.duplexgroup', package = 'paristools');
get_dg_coverage(
    read_duplexgroup(duplexgroup_file),
    sam_file \%>\% read_sam() \%>\% sam_to_loc_df() \%>\% cal_coverage()
)

}
