% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dg.R
\name{parse_group}
\alias{parse_group}
\title{parse a duplex group}
\usage{
parse_group(group)
}
\description{
parse a duplex group
}
\section{to do}{

here we parse each group separately to run faster, we can add tag to each group, and parse at once
}

