% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bed}
\alias{read_bed}
\title{read \code{.bed} file}
\usage{
read_bed(file, ...)
}
\arguments{
\item{file}{string. path to input file, passed onto \code{\link[readr:read_tsv]{readr::read_tsv()}}}

\item{...}{passed on to \code{\link[readr:read_tsv]{readr::read_tsv()}}}
}
\value{
a tibble of 6 variables
}
\description{
read \code{.bed} file
}
\examples{
sample_bed <- paste(
    'chr22\\t1000\\t5000\\tcloneA\\t960\\t+',
    'chr22\\t2000\\t6000\\tcloneB\\t900\\t-',
    sep = '\\n'
)
read_bed(sample_bed)

}
