% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dg.R
\name{read_duplexgroup}
\alias{read_duplexgroup}
\title{read \code{.duplexgroup} file}
\usage{
read_duplexgroup(file)
}
\arguments{
\item{file}{string. path to input file, passed onto \code{\link[readr:read_file]{readr::read_file()}}}
}
\value{
a tibble of 8 variables
}
\description{
read \code{.duplexgroup} file
}
\section{implementation}{

\code{read_duplexgroup()} runs quite fast, since we fully utilize R's
vectorisation feature, at the price of \emph{obscured} code.

\code{\link[=read_duplexgroup_old]{read_duplexgroup_old()}} is much clearer, it parses each group separately.
reading its source can help you understand the implementation.

In short, the most difficult part is, how to label each row with correct
identifier (group id here) after we concatenate each loc line and
\code{\link[=parse_locs]{parse_locs()}} at once,
}

\examples{
read_duplexgroup(system.file('extdata', 'Neat1_1.duplexgroup', package = 'paristools'))


}
