% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{read_sam}
\alias{read_sam}
\title{read \code{.sam} file}
\usage{
read_sam(file)
}
\arguments{
\item{file}{string. path to input file, passed onto \code{\link[readr:read_file]{readr::read_file()}}}
}
\value{
a tibble of 11 variables
}
\description{
We only preserve 11 mandatory fields
}
\details{
Since \code{.sam} file can cantain arbitrary columns after the required 11, and it doesn't contain a header line, I find it hard to use \code{\link[readr:read_tsv]{readr::read_tsv()}}. Finally, I use \code{\link[stringr:str_split_fixed]{stringr::str_split_fixed()}} to split the file into 12 columns, the last one which stored extra fields is discarded.
}
\examples{
read_sam(system.file('extdata', 'Neat1_1.Aligend_trunc.sam', package = 'paristools'))

}
