% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sam_to_loc_df}
\alias{sam_to_loc_df}
\title{find all mapped region in a \code{.sam} data.frame}
\arguments{
\item{sam}{data.frame. see \code{\link[=read_sam]{read_sam()}}}
}
\value{
a tibble of 4 variables, \code{chrom}, \code{strand}, \code{start}, \code{end}
}
\description{
find all mapped region in a \code{.sam} data.frame
}
\examples{
sam_file <- system.file('extdata', 'Neat1_1.Aligend_trunc.sam', package = 'paristools')
sam_file \%>\% read_sam() \%>\% sam_to_loc_df()

}
