% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_coverage_impl}
\alias{cal_coverage_impl}
\title{calculate genome coverage for one chromosome}
\arguments{
\item{loc_df}{data.frame. must contain 3 columns: \code{strand}, \code{start}, \code{end}}
}
\value{
list. two elements, \code{+} & \code{-}. each is a integer representing the genome, index is 1-based genome position, value is number of reads covering that position
}
\description{
calculate genome coverage for one chromosome
}
\details{
we use 1-based coordinate, \link{start, end}
}
\keyword{internal}
