% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dg.R
\name{parse_loc}
\alias{parse_loc}
\title{parse genome location specification in \code{.duplexgroup} file}
\usage{
parse_loc(loc)
}
\description{
parse genome location specification in \code{.duplexgroup} file
}
\details{
we assume chr only contains alpha-numeric characters
}
