% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{read_bed}
\alias{read_bed}
\title{read \code{.sam} file}
\usage{
read_bed(file, ...)
}
\arguments{
\item{file}{string. path to input file, passed onto \code{\link[readr:read_file]{readr::read_file()}}}

\item{...}{passed on to \code{\link[readr:read_tsv]{readr::read_tsv()}}}
}
\value{
a tibble of 6 variables
}
\description{
read \code{.sam} file
}
\examples{
NULL

}
